refactor: switch FFT/iFFT boundary to Float32Array and {real,imag} spectrum format
- fft::transform now accepts vector<float> directly, eliminating float->double->float round-trip
- addon.cpp: replace toDoubleVector/vectorToNumberArray with readFloat32Array/makeFloat32Array
- addon.cpp: complex spectrum boundary changed from array of {real,imag} objects to {real: Float32Array, imag: Float32Array}
- generateWindow, applyWindow, fftMagnitude, ifft all return Float32Array
- Commands.cpp, WavProcessor.cpp: remove redundant blockD double conversion
- JS tests updated for new Float32Array API

+++ src/audio/FFTProcessor.cpp
@@ -20,7 +20,7 @@ struct KissCfg {
 
 namespace fft {
 
-std::vector<std::complex<double>> transform(const std::vector<double> &input) {
+std::vector<std::complex<double>> transform(const std::vector<float> &input) {
     if (input.empty())
         throw dissonance::DspError("FFT input cannot be empty");
 
@@ -29,7 +29,7 @@ std::vector<std::complex<double>> transform(const std::vector<double> &input) {
 
     std::vector<kiss_fft_cpx> in(N), out(N);
     for (int i = 0; i < N; ++i) {
-        in[i].r = static_cast<float>(input[i]);
+        in[i].r = input[i];
         in[i].i = 0.0f;
     }
